backend/python/requirements.txt
tensorflow==2.15.0
grpcio==1.62.0
grpcio-tools==1.62.0
pytest>=7.4.0
numpy>=1.26.0
click>=8.1.0
rospy>=1.15.0
std-msgs>=0.5.0
python-dotenv>=1.0.0

